v1/requirements.txt
django
mysqlclient //pymysql
websocket-client
cryptography
djangorestframework djangorestframework-simplejwt
celery